+++ test/Codec/H264/H264InvalidSamplesParsing.cpp
@@ -40,7 +40,7 @@ void H264InvalidSamplesParsing::loadStream(const QString& szDirName, H264Stream&
 		QVERIFY(data.size() > 0);
 		bitstream.append(data);
 	}
-	QVERIFY(stream.parsePacket((uint8_t*)bitstream.data(), bitstream.size()) == expectedPacketParsingResult);
+	QVERIFY(stream.parsePacket((uint8_t*)bitstream.data(), (uint32_t)bitstream.size()) == expectedPacketParsingResult);
 }
 
 void H264InvalidSamplesParsing::test_h264EndOfStreamBitstream()
